Cum bugfix
The entire "Stat.prototype.DecreaseStat()" function was enclosed in a
"if(ideal)" statement, blocking any return statements if ideal was
false/0.
The function "Entity.prototype.OrgasmCum" relies on the output of
DecreaseStat and as such did not work properly.
Code that relies on the function:
"cumQ = balls.cum.DecreaseStat(0, cumQ, true);"